refactor: split multiarch login into another file
